% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseClassifierOneNN.R
\name{bClassifOneNN}
\alias{bClassifOneNN}
\title{1-NN classifier specification builder}
\usage{
bClassifOneNN()
}
\value{
A classifier specification built using
\code{\link{bClassif}} function.
}
\description{
Defines a 1-NN classifier specification for
use in conjunction with the semi-supervised classifiers in this package.
}

