% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervised.R
\name{checkProb}
\alias{checkProb}
\title{Check a matrix of probabilities}
\usage{
checkProb(prob, ninstances, classes)
}
\arguments{
\item{prob}{a probabilities matrix}

\item{ninstances}{expected number of rows in \code{prob}}

\item{classes}{expected columns names in \code{prob}}
}
\value{
the matrix \code{prob} with it columns in the order given by \code{classes}
}
\description{
Check the number of rows and the columns names
of a matrix of probabilities. If the columns are
unordered, they are ordered according to \code{classes}.
}
