% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoBC.R
\name{coBC}
\alias{coBC}
\title{Train the Co-bagging model}
\usage{
coBC(x, y, learner, learner.pars = list(), pred, pred.pars = list(),
  N = 3, perc.full = 0.7, u = 100, max.iter = 50)
}
\arguments{
\item{x}{A matrix or a dataframe with the training instances.}

\item{y}{A vector with the labels of training instances. In this vector the unlabeled instances
are specified with the value \code{NA}.}

\item{learner}{either a function or a string naming the function for 
training a supervised base classifier}

\item{learner.pars}{A list with parameters that are to be passed to the \code{learner}
function at each Co-bagging iteration.}

\item{pred}{either a function or a string naming the function for
predicting the probabilities per classes,
using a base classifier trained with function \code{learner}.}

\item{pred.pars}{A list with parameters that are to be passed to the \code{pred}
function.}

\item{N}{The number of classifiers used as committee members. Default is 3.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the Co-bagging process is stopped.
Default is 0.7.}

\item{u}{Number of unlabeled instances in the pool. Default is 100.}

\item{max.iter}{Maximum number of iterations to execute the self-labeling process. 
Default is 50.}
}
\description{
Trains a model for classification,
according to Co-bagging algorithm.
}
