% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{Time series data set}
\format{A data frame with 56 rows and 287 variables including the class.}
\source{
\url{www.cs.ucr.edu/~eamonn/time_series_data/}
}
\usage{
data(coffee)
}
\description{
A dataset containing 56 times series z-normalized. Time series length is 286.
}
\keyword{datasets}

