% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Democratic.R
\name{democratic}
\alias{democratic}
\title{Train a Democratic model}
\usage{
democratic(x, y, learners, learners.pars, preds, preds.pars)
}
\arguments{
\item{x}{A matrix or a dataframe with the training instances.}

\item{y}{A vector with the labels of training instances. In this vector the unlabeled instances
are specified with the value \code{NA}.}

\item{learners}{A list of learner functions from diferents learning squemes.
The learners are used to train supervised base classifiers.}

\item{learners.pars}{A list of parameter sets, 
where each set is used with its corresponding learner,
according to the order in \code{learners}.}

\item{preds}{A list of prediction functions to obtain the probabilities per classes.
Each function is used in conjuntion with the trained model provided in \code{learners}.}

\item{preds.pars}{A list of parameter sets for the functions in \code{preds} list. 
According to the position, each parameter set corresponds with a specific function.}
}
\value{
The trained model.
}
\description{
Trains a model for classification,
according to Democratic algorithm.
}
