% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervised.R
\name{getClass}
\alias{getClass}
\title{Get classes from a matrix of probabilities}
\usage{
getClass(prob, ninstances, classes)
}
\arguments{
\item{prob}{a probabilities matrix}

\item{ninstances}{expected number of rows in \code{prob}}

\item{classes}{expected columns names in \code{prob}}
}
\value{
a factor with classes
}
