% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervised.R
\name{getClassIdx}
\alias{getClassIdx}
\title{Get classes from a matrix of probabilities and 
return the classes indexes}
\usage{
getClassIdx(prob)
}
\arguments{
\item{prob}{a probabilities matrix}
}
\value{
a vector of indexes corresponding to \code{classes}
}
