% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervised.R
\name{getProb}
\alias{getProb}
\title{Check and order a matrix of probabilities}
\usage{
getProb(prob, ninstances, classes)
}
\arguments{
\item{prob}{a probabilities matrix}

\item{ninstances}{expected number of rows in \code{prob}}

\item{classes}{expected columns names in \code{prob}}
}
\value{
the matrix \code{prob} with it columns in the order given by \code{classes}
}
