% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseClassifierOneNN.R
\name{oneNN}
\alias{oneNN}
\title{1-NN supervised classifier builder}
\usage{
oneNN(x = NULL, y)
}
\arguments{
\item{x}{This argument is not used, the reason why he gets is to fulfill an agreement}

\item{y}{a vector with the labels of training instances}
}
\value{
A model wish the data needed to use 1-NN
}
\description{
Build a model using the given data to be capable
of predict the label or the probabilities of other instances,
according to 1-NN algorithm.
}

