% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETRED.R
\name{predict.setred}
\alias{predict.setred}
\title{Predictions of the SETRED method}
\usage{
\method{predict}{setred}(object, x, ...)
}
\arguments{
\item{object}{SETRED model built with the \code{\link{setred}} function.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix 
with the distances between the unseen instances and the selected training instances, 
or a matrix of instances.}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances according to the \code{setred} model.
}
\details{
For additional help see \code{\link{setred}} examples.
}
