% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNNRCE.R
\name{predict.snnrce}
\alias{predict.snnrce}
\title{Predictions of the SNNRCE method}
\usage{
\method{predict}{snnrce}(object, x, ...)
}
\arguments{
\item{object}{SNNRCE model built with the \code{\link{snnrce}} function.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix 
with the distances between the unseen instances and the selected training instances, 
or a matrix of instances.}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances according to the \code{snnrce} model.
}
\details{
For additional help see \code{\link{snnrce}} examples.
}
