% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfTraining.R
\name{selfTraining}
\alias{selfTraining}
\title{Train a Self-training model}
\usage{
selfTraining(x, y, learner, learner.pars = list(), pred, pred.pars = list(),
  max.iter = 50, perc.full = 0.7, thr.conf = 0.5, train.model = TRUE)
}
\arguments{
\item{x}{A matrix or a dataframe with the training instances.}

\item{y}{A vector with the labels of training instances. In this vector the unlabeled instances
are specified with the value \code{NA}.}

\item{learner}{either a function or a string naming the function for 
training a supervised base classifier}

\item{learner.pars}{A list with parameters that are to be passed to the \code{learner}
function at each self-training iteration.}

\item{pred}{either a function or a string naming the function for
predicting the probabilities per classes,
using a base classifier trained with function \code{learner}.}

\item{pred.pars}{A list with parameters that are to be passed to the \code{pred}
function.}

\item{max.iter}{Maximum number of iterations to execute the self-labeling process. 
Default is 50.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the self-training process is stopped.
Default is 0.7.}

\item{thr.conf}{A number between 0 and 1 that indicates the confidence theshold.
At each iteration, only the new label examples with a confidence greater than 
this value (\code{thr.conf}) are added to training set.}
}
\description{
Trains a model for classification,
according to Self-training algorithm.
}
