% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfTraining.R
\name{selfTrainingBase}
\alias{selfTrainingBase}
\title{Train a Self-training model}
\usage{
selfTrainingBase(y, learnerB, predB, max.iter = 50, perc.full = 0.7,
  thr.conf = 0.5)
}
\arguments{
\item{y}{A vector with the labels of training instances. In this vector the unlabeled instances
are specified with the value \code{NA}.}

\item{learnerB}{a function for training a supervised base classifier
This function must be two parameters, indexes and cls, where indexes
indicate the instances to use and cls specify the classes of those instances.}

\item{predB}{a function for predicting the probabilities per classes.
This function must be two parameters, model and indexes, where model
provide the classifier trained with function \code{learner} and
indexes indicate the instances to predict.}

\item{max.iter}{Maximum number of iterations to execute the self-labeling process. 
Default is 50.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the self-training process is stopped.
Default is 0.7.}

\item{thr.conf}{A number between 0 and 1 that indicates the theshold confidence.
At each iteration, only the new label examples with a confidence greater than 
this value (\code{thr.conf}) are added to training set.}
}
\value{
The model trained with \code{learnerB} for the labeled set obtained.
}
\description{
Trains a model for classification,
according to Self-training algorithm.
}
