% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETRED.R
\name{setredBase}
\alias{setredBase}
\title{SETRED base method}
\usage{
setredBase(y, D, learnerB, predB, theta = 0.1, max.iter = 50,
  perc.full = 0.7)
}
\arguments{
\item{y}{A vector with the labels of training instances. In this vector the 
unlabeled instances are specified with the value \code{NA}.}

\item{D}{A distance matrix between all the training instances. This matrix is used to 
construct the neighborhood graph.}

\item{learnerB}{A function for training a supervised base classifier.
This function needs two parameters, indexes and cls, where indexes indicates
the instances to use and cls specifies the classes of those instances.}

\item{predB}{A function for predicting the probabilities per classes.
This function must be two parameters, model and indexes, where the model
is a classifier trained with \code{learnerB} function and
indexes indicates the instances to predict.}

\item{theta}{Rejection threshold to test the critical region. Default is 0.1.}

\item{max.iter}{Maximum number of iterations to execute the self-labeling process. 
Default is 50.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the self-training process is stopped.
Default is 0.7.}
}
\value{
A list object of class "setredBase" containing:
\describe{
  \item{model}{The final base classifier trained using the enlarged labeled set.}
  \item{included.insts}{The indexes of the training instances used to 
  train the \code{model}. These indexes include the initial labeled instances
  and the newly labeled instances.
  Those indexes are relative to \code{y} argument.}
}
}
\description{
SETRED is a variant of the self-training classification method 
(\code{\link{selfTraining}}) with a different addition mechanism. 
The SETRED classifier is initially trained with a 
reduced set of labeled examples. Then it is iteratively retrained with its own most 
confident predictions over the unlabeled examples. SETRED uses an amending scheme 
to avoid the introduction of noisy examples into the enlarged labeled set. For each 
iteration, the mislabeled examples are identified using the local information provided 
by the neighborhood graph.
}
\details{
SetredBase can be helpful in those cases where the method selected as 
base classifier needs a \code{learner} and \code{pred} functions with other
specifications. For more information about the general setred method,
please see \code{\link{setred}} function. Essentially, \code{setred}
function is a wrapper of \code{setredBase} function.
}
