% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNNRCE.R
\name{snnrce}
\alias{snnrce}
\title{Train a SNNRCE model}
\usage{
snnrce(x, y, dist = "Euclidean", alpha = 0.1)
}
\arguments{
\item{x}{A matrix or a dataframe with the training instances.}

\item{y}{A vector with the labels of training instances. 
In this vector the unlabeled instances are specified with the value \code{NA}.}

\item{dist}{The name of a distance method available in \code{proxy} package or
a function defined by the user that computes the distance between two instances.
Default is Euclidean distance.}

\item{alpha}{Rejection threshold to test the critical region. Default is 0.1.}
}
\value{
The trained model.
}
\description{
Trains a model for classification,
according to SNNRCE algorithm.
}
