% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNNRCE.R
\name{snnrceBase}
\alias{snnrceBase}
\title{SNNRCE base method}
\usage{
snnrceBase(D, y, alpha = 0.1)
}
\arguments{
\item{D}{A distance matrix between all the training instances. This matrix is used to 
construct the neighborhood graph.}

\item{y}{A vector with the labels of training instances. In this vector the 
unlabeled instances are specified with the value \code{NA}.}

\item{alpha}{Rejection threshold to test the critical region. Default is 0.1.}
}
\value{
A list object of class "snnrceBase" containing:
\describe{
  \item{model}{The final base classifier trained using the enlarged labeled set.}
  \item{included.insts}{The indexes of the training instances used to 
  train the \code{model}. These indexes include the initial labeled instances
  and the newly labeled instances.
  Those indexes are relative to \code{y} argument.}
}
}
\description{
SNNRCE is a variant of the self-training classification 
method (\code{\link{selfTraining}}) with a different 
addition mechanism and a fixed learning scheme (1-NN). SNNRCE uses an amending scheme 
to avoid the introduction of noisy examples into the enlarged labeled set.
The mislabeled examples are identified using the local information provided 
by the neighborhood graph. A statistical test using cut edge weight is used to modify 
the labels of the missclassified examples.
}
