% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{statistics}
\alias{statistics}
\title{Statistics calculation}
\usage{
statistics(predicted, real)
}
\arguments{
\item{predicted}{a vector with the predicted labels}

\item{real}{a vector with the real labels}
}
\value{
a named list with the statistics values calculated
}
\description{
Calculates the statistics:
Accuracy, Cohen's kappa and F-measure
}

