% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriTraining.R
\name{triTraining}
\alias{triTraining}
\title{Train a Tri-training model}
\usage{
triTraining(x, y, learner, learner.pars = list(), pred, pred.pars = list())
}
\arguments{
\item{x}{A matrix or a dataframe with the training instances.}

\item{y}{A vector with the labels of training instances. In this vector the unlabeled instances
are specified with the value \code{NA}.}

\item{learner}{either a function or a string naming the function for 
training a supervised base classifier}

\item{learner.pars}{A list with parameters that are to be passed to the \code{learner}
function.}

\item{pred}{either a function or a string naming the function for
predicting the probabilities per classes,
using a base classifier trained with the \code{learner} function.}

\item{pred.pars}{A list with parameters that are to be passed to the \code{pred}
function.}
}
\value{
The trained model.
}
\description{
Trains a model for classification,
according to Tri-training algorithm.
}
