% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriTraining.R
\name{triTrainingBase}
\alias{triTrainingBase}
\title{Tri-training base method}
\usage{
triTrainingBase(y, learnerB, predB)
}
\arguments{
\item{y}{A vector with the labels of training instances. In this vector the 
unlabeled instances are specified with the value \code{NA}.}

\item{learnerB}{A function for training three supervised base classifiers.
This function needs two parameters, indexes and cls, where indexes indicates
the instances to use and cls specifies the classes of those instances.}

\item{predB}{A function for predicting the probabilities per classes.
This function must be two parameters, model and indexes, where the model
is a classifier trained with \code{learnerB} function and
indexes indicates the instances to predict.}
}
\value{
A list object of class "triTrainingBase" containing:
\describe{
  \item{model}{The final base classifier trained using the enlarged labeled set.}
  \item{included.insts}{The indexes of the training instances used to 
  train the \code{model}. These indexes include the initial labeled instances
  and the newly labeled instances.
  Those indexes are relative to \code{y} argument.}
}
}
\description{
Tri-training is a semi-supervised learning algorithm with a co-training 
style. This algorithm trains three classifiers with the same learning scheme from a 
reduced set of labeled examples. For each iteration, an unlabeled example is labeled 
for a classifier if the other two classifiers agree on the labeling proposed.
}
\details{
TriTrainingBase can be helpful in those cases where the method selected as 
base classifier needs a \code{learner} and \code{pred} functions with other
specifications. For more information about the general triTraining method,
please see \code{\link{triTraining}} function. Essentially, \code{triTraining}
function is a wrapper of \code{triTrainingBase} function.
}
