% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriTraining.R
\name{triTrainingCombine}
\alias{triTrainingCombine}
\title{Combining the hypothesis}
\usage{
triTrainingCombine(pred)
}
\arguments{
\item{pred}{A list with the predictions of each classifiers}
}
\value{
A vector of classes
}
\description{
This function combines the predictions obtained 
by the set of classifiers.
}
