% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoBC.R
\name{coBC}
\alias{coBC}
\title{CoBC method}
\usage{
coBC(x, y, learner, learner.pars = list(), pred, pred.pars = list(),
  x.dist = FALSE, N = 3, perc.full = 0.7, u = 100, max.iter = 50)
}
\arguments{
\item{x}{A object that can be coerced as matrix. This object has two possible 
interpretations according to the value set in \code{x.dist} argument: 
a matrix distance between the training examples or a matrix with the 
training instances where each row represents a single instance.}

\item{y}{A vector with the labels of the training instances. In this vector 
the unlabeled instances are specified with the value \code{NA}.}

\item{learner}{either a function or a string naming the function for 
training the supervised base classifiers.}

\item{learner.pars}{A list with additional parameters for the
\code{learner} function if necessary.}

\item{pred}{either a function or a string naming the function for
predicting the probabilities per classes,
using the base classifiers trained with the \code{learner} function.}

\item{pred.pars}{A list with additional parameters for the
\code{pred} function if necessary.}

\item{N}{The number of classifiers used as committee members. All these classifiers 
are trained using the \code{learnerB} function. Default is 3.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the self-labeling process is stopped.
Default is 0.7.}

\item{u}{Number of unlabeled instances in the pool. Default is 100.}

\item{max.iter}{Maximum number of iterations to execute in the self-labeling process. 
Default is 50.}
}
\value{
A list object of class "coBC" containing:
\describe{
  \item{models}{The final three base classifiers trained using the enlarged labeled set.}
  \item{included.insts}{The indexes of the total of training instances used to 
  train the three \code{models}. These indexes include the initial labeled instances
  and the newly labeled instances.
  These indexes are relative to the \code{y} argument.}
  \item{indexes}{List of three vectors of indexes related to the training instances 
  used per each classifier. These indexes are relative to \code{included.insts}.}
  \item{classes}{The levels of \code{y} factor.}
  \item{pred}{The function provided in the \code{pred} argument.}
  \item{pred.pars}{The list provided in the \code{pred.pars} argument.}
  \item{x.dist}{The value provided in the \code{x.dist} argument.}
}
}
\description{
Co-Training by Committee (CoBC) is a semi-supervised learning algorithm 
with a co-training style. This algorithm trains \code{N} classifiers with the learning 
scheme defined in \code{learner} argument using a reduced set of labeled examples. For 
each iteration, an unlabeled 
example is labeled for a classifier if the most confident classifications assigned by the 
other \code{N-1} classifiers agree on the labeling proposed. The unlabeled examples 
candidates are selected randomly from a pool of size \code{u}.
}
\details{
This method trains an ensemble of diverse classifiers. To promote the initial diversity 
the classifiers are trained from the reduced set of labeled examples by Bagging.
The stopping criterion is defined through the fulfillment of one of the following
criteria: the algorithm reaches the number of iterations defined in \code{max.iter}
parameter or the portion of unlabeled set, defined in \code{perc.full} parameter,
is moved to the enlarged labeled set of the classifiers.
}
