% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoBC.R
\name{coBCBase}
\alias{coBCBase}
\title{CoBC base method}
\usage{
coBCBase(y, learnerB, predB, N = 3, perc.full = 0.7, u = 100,
  max.iter = 50)
}
\arguments{
\item{y}{A vector with the labels of training instances. In this vector the 
unlabeled instances are specified with the value \code{NA}.}

\item{learnerB}{A function for training \code{N} supervised base classifiers.
This function needs two parameters, indexes and cls, where indexes indicates
the instances to use and cls specifies the classes of those instances.}

\item{predB}{A function for predicting the probabilities per classes.
This function must be two parameters, model and indexes, where the model
is a classifier trained with \code{learnerB} function and
indexes indicates the instances to predict.}

\item{N}{The number of classifiers used as committee members. All these classifiers 
are trained using the \code{learnerB} function. Default is 3.}

\item{perc.full}{A number between 0 and 1. If the percentage 
of new labeled examples reaches this value the self-labeling process is stopped.
Default is 0.7.}

\item{u}{Number of unlabeled instances in the pool. Default is 100.}

\item{max.iter}{Maximum number of iterations to execute in the self-labeling process. 
Default is 50.}
}
\value{
A list object of class "coBCBase" containing:
\describe{
  \item{model}{The final three base classifiers trained using the enlarged labeled set.}
  \item{instances.index}{The indexes of the total of training instances used to 
  train the three models. These indexes include the initial labeled instances
  and the newly labeled instances.
  These indexes are relative to the \code{y} argument.}
  \item{model.index}{List of three vectors of indexes related to the training instances 
  used per each classifier. These indexes are relative to \code{instances.index}.}
  \item{classes}{The levels of \code{y} factor.}
}
}
\description{
CoBC is a semi-supervised learning algorithm with a co-training 
style. This algorithm trains \code{N} classifiers with the learning scheme defined in 
\code{learnerB} using a reduced set of labeled examples. For each iteration, an unlabeled 
example is labeled for a classifier if the most confident classifications assigned by the 
other \code{N-1} classifiers agree on the labeling proposed. The unlabeled examples 
candidates are selected randomly from a pool of size \code{u}.
}
\details{
coBCBase can be helpful in those cases where the method selected as 
base classifier needs a \code{learner} and \code{pred} functions with other
specifications. For more information about the general coBC method,
please see \code{\link{coBC}} function. Essentially, \code{coBC}
function is a wrapper of \code{coBCBase} function.
}
