% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoBC.R
\name{coBCCombine}
\alias{coBCCombine}
\title{Combining the hypothesis}
\usage{
coBCCombine(h.prob, classes)
}
\arguments{
\item{h.prob}{A list of probability matrices.}

\item{classes}{The classes in the same order that appear 
in the columns of each matrix in \code{h.prob}.}
}
\value{
A probability matrix
}
\description{
This function combines the probabilities predicted by the committee of 
classifiers.
}
