% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfTraining.R
\name{predict.selfTraining}
\alias{predict.selfTraining}
\title{Predictions of the Self-training method}
\usage{
\method{predict}{selfTraining}(object, x, ...)
}
\arguments{
\item{object}{Self-training model built with the \code{\link{selfTraining}} function.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix 
with the distances between the unseen instances and the selected training instances, 
or a matrix of instances.}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances according to the \code{selfTraining} model.
}
\details{
For additional help see \code{\link{selfTraining}} examples.
}
