% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETRED.R
\name{predict.setred}
\alias{predict.setred}
\title{Model Predictions}
\usage{
\method{predict}{setred}(object, x, ...)
}
\arguments{
\item{object}{SETRED model object built with \code{\link{setred}} function.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix
with the distances between the unseen instances and the selected training instances,
or a matrix of instances.}

\item{...}{Additional objects that can be coerced as matrix.
Depending on how was the model built, this objects are used or ignored.
This objects are interpreted as matrices of distances between the unseen instances
and the selected training instances.}
}
\value{
Vector with labels of instances.
}
\description{
Predicts the label of instances according to SETRED model.
}
\details{
For additional help see \code{\link{setred}} examples.
}

