% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriTraining.R
\name{predict.triTraining}
\alias{predict.triTraining}
\title{Predictions of the Tri-training method}
\usage{
\method{predict}{triTraining}(object, x, ...)
}
\arguments{
\item{object}{Tri-training model built with the \code{\link{triTraining}} function.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix 
with the distances between the unseen instances and the selected training instances, 
or a matrix of instances.}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances according to the \code{triTraining} model.
}
\details{
For additional help see \code{\link{triTraining}} examples.
}
