% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriTraining.R
\name{triTraining}
\alias{triTraining}
\title{Tri-training method}
\usage{
triTraining(x, y, learner, learner.pars = list(), pred, pred.pars = list(),
  x.dist = FALSE)
}
\arguments{
\item{x}{A object that can be coerced as matrix. This object has two possible 
interpretations according to the value set in the \code{x.dist} argument:
a matrix distance between the training examples or a matrix with the 
training instances where each row represents a single instance.}

\item{y}{A vector with the labels of the training instances. In this vector 
the unlabeled instances are specified with the value \code{NA}.}

\item{learner}{either a function or a string naming the function for 
training the supervised base classifiers.}

\item{learner.pars}{A list with additional parameters for the
\code{learner} function if necessary.}

\item{pred}{either a function or a string naming the function for
predicting the probabilities per classes,
using the base classifiers trained with the \code{learner} function.}

\item{pred.pars}{A list with additional parameters for the
\code{pred} function if necessary.}

\item{x.dist}{A boolean value that indicates if \code{x} is or not a distance matrix.
Default is \code{FALSE}.}
}
\value{
A list object of class "triTraining" containing:
\describe{
  \item{model}{The final three base classifiers trained using the enlarged labeled set.}
  \item{instances.index}{The indexes of the total of training instances used to 
  train the three models. These indexes include the initial labeled instances
  and the newly labeled instances.
  These indexes are relative to the \code{y} argument.}
  \item{model.index}{List of three vectors of indexes related to the training instances 
  used per each classifier. These indexes are relative to \code{instances.index}.}
  \item{classes}{The levels of \code{y} factor.}
  \item{pred}{The function provided in the \code{pred} argument.}
  \item{pred.pars}{The list provided in the \code{pred.pars} argument.}
  \item{x.dist}{The value provided in the \code{x.dist} argument.}
}
}
\description{
Tri-training is a semi-supervised learning algorithm with a co-training 
style. This algorithm trains three classifiers with the same learning scheme from a 
reduced set of labeled examples. For each iteration, an unlabeled example is labeled 
for a classifier if the other two classifiers agree on the labeling proposed.
}
\details{
Tri-training initiates the self-labeling process by training three models from the 
original labeled set, using the \code{learner} function specified. 
In each iteration, the algorithm detects unlabeled examples on which two classifiers 
agree with the classification and includes these instances in the enlarged set of the 
third classifier under certain conditions. The generation of the final hypothesis is 
produced via the majority voting. The iteration process ends when no changes occurs in 
any model during a complete iteration.
}
\references{
ZhiHua Zhou and Ming Li.\cr
\emph{Tri-training: exploiting unlabeled data using three classifiers.}\cr
IEEE Transactions on Knowledge and Data Engineering, 17(11):1529-1541, Nov 2005. ISSN 1041-4347. doi: 10.1109/TKDE.2005. 186.
}
